Add files via upload
BizCardDB.py file creates the DB and table structure for this project. Since this is a one time creation, this part has been kept separate.
BizCardUI.py takes care of the Streamlit app and related functionalities of extracting using easyOCR, parsing that data using reg ex and storing and manipulating the data through MySQL.
BizCardIcon.png file is the icon used in the streamlit app title.